More test problems - this time because of mock-overuse

+++ kalite/contentload/management/commands/unpack_assessment_zip.py
@@ -107,6 +107,10 @@ def unpack_zipfile_to_content_folder(zf):
 
     ensure_dir(folder)
     zf.extractall(folder)
+
+    # If assessmentitems.version exists, copy it to another location outside
+    # of the channel folder because for some reason a test expects it to be
+    # there.
     version_file = os.path.join(folder, 'assessmentitems.version')
     version_file_copied_dest = os.path.join(
         settings.ASSESSMENT_ITEM_ROOT,
@@ -114,11 +118,14 @@ def unpack_zipfile_to_content_folder(zf):
     )
     if os.path.isfile(version_file_copied_dest):
         os.unlink(version_file_copied_dest)
-    # Ensure that special files are in their configured locations
-    shutil.copy(
-        version_file,
-        version_file_copied_dest
-    )
+    # Test that file exists because there's a test that mocks unzipping and
+    # then this would fail because a file that should exist doesn't (doh)
+    if os.path.isfile(version_file):
+        # Ensure that special files are in their configured locations
+        shutil.copy(
+            version_file,
+            version_file_copied_dest
+        )
 
 
 def is_valid_url(url):
